\name{IGM.plot.ns}
\alias{IGM.plot.ns}
\title{
Generic method for plotting network spikes.
}
\description{
The generic plotting function for network spikes.
}
\usage{
IGM.plot.ns(ns, ...)
}
\arguments{
  \item{ns}{
    The network spike object returned after running calculate.network.spikes.
}
  \item{\dots}{
  Additional plotting options that is general to plot functions.
}
}

\value{
None. network spikes related plots will be generated in current plotting device.
}

\keyword{network spikes}
\keyword{plot}
